delete function module
